+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Defining a contact point for a dataset is done throught a form field
 - Allow wildcards in redirect_uris for Oauth2Client [#2935](https://github.com/opendatateam/udata/pull/2935)
 - Allow for being one day late on update fulfilled in time [#2941](https://github.com/opendatateam/udata/pull/2941)
+- When a topic is deleted, corresponding discussions are purged [#2944](https://github.com/opendatateam/udata/pull/2944)
 
 ## 7.0.1 (2023-12-06)